webdriver: trying out execute script
An example of webdriver executeScript.
